Production support
I have added "dist" script for npm start. It allows us to build project's installer.

While build was successful at first, I was unable to load some of the resources due to the build's storing.

I had to learn a bit about resources/app.asar

This is also why there are changes made to the hyperlinks and other source like links.

I also decided to organise links with path.join() as it formats paths adequately. I no longer have to deal with mixed slashes, plus, it makes it easier to read and understand the code thanks to its semantic like format.

I also played a little bit with an a npm package "node-notifier". It helped me with testing as I was unable to log my mistakes in production. I used it as a mental breakpoint while doing the dry run tests and thought of it's inclusion due to its usefulness.